relocate some internal functions from commands.py to internal.py
