+++ pybatfish/client/commands.py
@@ -27,7 +27,6 @@ import six
 from requests import HTTPError
 
 from pybatfish.client.consts import CoordConsts, WorkStatusCode
-from pybatfish.datamodel import answer
 from pybatfish.datamodel.primitives import (  # noqa: F401
     Edge, Interface)
 from pybatfish.datamodel.referencelibrary import (NodeRoleDimension,
@@ -39,7 +38,7 @@ from pybatfish.util import (BfJsonEncoder, get_uuid, validate_name, zip_dir)
 from . import resthelper, restv2helper, workhelper
 from .options import Options
 from .session import Session
-from .workhelper import (_get_data_get_question_templates, get_work_status,
+from .workhelper import (get_work_status,
                          kill_work)
 
 # TODO: normally libraries don't configure logging in code
@@ -138,44 +137,6 @@ def bf_add_reference_book(book):
     :type book: :class:`pybatfish.datamodel.referencelibrary.ReferenceBook`
     """
     restv2helper.add_reference_book(bf_session, book)
-
-
-def _bf_answer_obj(question_str, parameters_str, question_name,
-                   background, snapshot, reference_snapshot):
-    # type: (str, str, str, bool, str, Optional[str]) -> Union[str, Dict]
-
-    json.loads(parameters_str)  # a syntactic check for parametersStr
-    if not question_name:
-        question_name = Options.default_question_prefix + "_" + get_uuid()
-
-    # Upload the question
-    json_data = workhelper.get_data_upload_question(bf_session, question_name,
-                                                    question_str,
-                                                    parameters_str)
-    resthelper.get_json_response(bf_session,
-                                 CoordConsts.SVC_RSC_UPLOAD_QUESTION, json_data)
-
-    # Answer the question
-    work_item = workhelper.get_workitem_answer(bf_session, question_name,
-                                               snapshot, reference_snapshot)
-    workhelper.execute(work_item, bf_session, background)
-
-    if background:
-        return work_item.id
-
-    # get the answer
-    answer_bytes = resthelper.get_answer(bf_session, snapshot, question_name,
-                                         reference_snapshot)
-
-    # In Python 3.x, answer needs to be decoded before it can be used
-    # for things like json.loads (<= 3.6).
-    if six.PY3:
-        answer_string = answer_bytes.decode(encoding="utf-8")
-    else:
-        answer_string = answer_bytes
-    answer_obj = json.loads(answer_string)
-
-    return answer.from_string(answer_obj[1]['answer'])
 
 
 def bf_auto_complete(completionType, query, maxSuggestions=None):
@@ -613,14 +574,6 @@ def bf_list_snapshots(verbose=False):
         snapshots and metadata (if `verbose=True`)
     """
     return restv2helper.list_snapshots(bf_session, verbose)
-
-
-def _bf_get_question_templates():
-    jsonData = _get_data_get_question_templates(bf_session)
-    jsonResponse = resthelper.get_json_response(bf_session,
-                                                CoordConsts.SVC_RSC_GET_QUESTION_TEMPLATES,
-                                                jsonData)
-    return jsonResponse[CoordConsts.SVC_KEY_QUESTION_LIST]
 
 
 def bf_put_node_roles(node_roles_data):
